+++ src/Tests/RoleKeysManagement.php
@@ -1,13 +1,7 @@
 <?php
 
-/**
- * @file
- * Definition of Drupal\password_policy\Tests\PasswordManualReset.
- */
-
 namespace Drupal\pubkey_encrypt\Tests;
 
-use Drupal\simpletest\WebTestBase;
 use Drupal\user\Entity\Role;
 
 /**
@@ -15,26 +9,7 @@ use Drupal\user\Entity\Role;
  *
  * @group pubkey_encrypt
  */
-class RoleKeysManagement extends WebTestBase {
-
-  public static $modules = array(
-    'key',
-    'encrypt',
-    'pubkey_encrypt',
-
-  );
-
-  protected $profile = 'minimal';
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function setUp() {
-    parent::setUp();
-
-    // Have the module initialized.
-    $this->initializePubkeyEncrypt();
-  }
+class RoleKeysManagement extends PubkeyEncryptTestBase {
 
   /**
    * Test Role keys.
@@ -113,19 +88,6 @@ class RoleKeysManagement extends WebTestBase {
     // Test that the Role key has been deleted.
     $new_role_key = $key_repository->getKey($new_role_id . "_role_key");
     $this->assertNull($new_role_key, "Role key gets deleted upon the deletion of a role");
-  }
-
-  protected function initializePubkeyEncrypt(){
-    $config = \Drupal::service('config.factory')
-      ->getEditable('pubkey_encrypt.initialization_settings');
-    $config->set('module_initialized', 1)
-      // Use default plugins provided by the module during initialization.
-      ->set('asymmetric_keys_generator', 'openssl_default')
-      ->set('asymmetric_keys_generator_configuration', array('key_size' => '2048'))
-      ->set('login_credentials_provider', 'user_passwords')
-      ->save();
-    \Drupal::service('pubkey_encrypt.pubkey_encrypt_manager')
-      ->initializeModule();
   }
 
 }
